+++ Deeploy/Targets/XDNA2/Bindings.py
@@ -5,7 +5,7 @@
 from Deeploy.AbstractDataTypes import PointerClass
 from Deeploy.CommonExtensions.DataTypes import bfloat16_t
 from Deeploy.DeeployTypes import NodeBinding
-from Deeploy.MLIRDataTypes import MLIRCodeTransformation
+from Deeploy.MLIRAIETypes import MLIRCodeTransformation
 from Deeploy.Targets.XDNA2.CodeTransformationPasses.MLIRComputeCorePass import MLIRComputeCorePass
 from Deeploy.Targets.XDNA2.CodeTransformationPasses.MLIRObjectFifoPass import MLIRObjectFifoPass
 from Deeploy.Targets.XDNA2.CodeTransformationPasses.MLIRRuntimeSequencePass import MLIRRuntimeSequencePass
